[niconico] Separate heartbeat mechanism.
